mac/requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
av==15.0.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
#-e git+https://github.com/rednote-hilab/dots.ocr@6a5e2ba95c4cbcffcbd11cad2eb4e42c4e23f952#egg=dots_ocr
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fsspec==2025.7.0
gradio==5.41.0
gradio_client==1.11.0
gradio_image_annotation==0.4.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.28.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openai==1.99.1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2025.7.34
requests==2.32.4
rich==14.1.0
ruff==0.12.7
safehttpx==0.1.6
safetensors==0.6.1
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.51.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
